add spring security authentication to market-app controllers

+++ market-app/src/main/java/ru/yandex/practicum/mymarket/service/impl/OrderProcessingServiceImpl.java
@@ -54,7 +54,7 @@ public class OrderProcessingServiceImpl implements OrderProcessingService {
                     return paymentsApi.processPayment(
                                     new PaymentRequest()
                                             .userId(userId)
-                                            .amount(total.doubleValue())
+                                            .amount(total)
                             )
                             .flatMap(paymentResult ->
                                     orderService.create(cartItemDtos, userId)
